STM32 FLASH fixes: use size_t instead of uint16_t, make interface more generic.  From Freddie Chopin

+++ system/free/free.c
@@ -51,9 +51,9 @@
 
 static void free_getprogmeminfo(struct mallinfo * mem)
 {
-  uint16_t page = 0, stpage = 0xFFFF;
-  uint16_t pagesize = 0;
-  int status;
+  size_t page = 0, stpage = 0xFFFF;
+  size_t pagesize = 0;
+  ssize_t status;
 
   mem->arena    = 0;
   mem->fordblks = 0;
